Cleanup HelloWorld Example
